Moved AE likelihood tracker to its own file, removed from private Utils file
# Conflicts:
#	Examples/BeeTrackingTool/main.swift
