+++ Examples/OISTVisualizationTool/Util.swift
@@ -43,63 +43,6 @@ public func makeOISTTrainingBatch(dataset: OISTBeeVideo, appearanceModelSize: (I
 }
 
 
-/// Returns a tracking configuration for a tracker using an RAE.
-///
-/// Parameter model: The RAE model to use.
-/// Parameter statistics: Normalization statistics for the frames.
-/// Parameter frames: The frames of the video where we want to run tracking.
-/// Parameter targetSize: The size of the target in the frames.
-public func makeNaiveBayesRAETracker(
-  model: DenseRAE,
-  statistics: FrameStatistics,
-  frames: [Tensor<Float>],
-  targetSize: (Int, Int),
-  foregroundModel: MultivariateGaussian,
-  backgroundModel: GaussianNB
-) -> TrackingConfiguration<Tuple1<Pose2>> {
-  var variableTemplate = VariableAssignments()
-  var frameVariableIDs = [Tuple1<TypedID<Pose2>>]()
-  for _ in 0..<frames.count {
-    frameVariableIDs.append(
-      Tuple1(
-        variableTemplate.store(Pose2())
-        ))
-  }
-  return TrackingConfiguration(
-    frames: frames,
-    variableTemplate: variableTemplate,
-    frameVariableIDs: frameVariableIDs,
-    addPriorFactor: { (variables, values, graph) -> () in
-      let (poseID) = unpack(variables)
-      let (pose) = unpack(values)
-      graph.store(WeightedPriorFactor(poseID, pose, weight: 1e-1))
-    },
-    addTrackingFactor: { (variables, frame, graph) -> () in
-      let (poseID) = unpack(variables)
-      graph.store(
-        ProbablisticTrackingFactor(poseID,
-          measurement: statistics.normalized(frame),
-          encoder: model,
-          patchSize: targetSize,
-          appearanceModelSize: targetSize,
-          foregroundModel: foregroundModel,
-          backgroundModel: backgroundModel,
-          maxPossibleNegativity: 1e1
-        )
-      )
-    },
-    addBetweenFactor: { (variables1, variables2, graph) -> () in
-      let (poseID1) = unpack(variables1)
-      let (poseID2) = unpack(variables2)
-      graph.store(WeightedBetweenFactorPose2(poseID1, poseID2, Pose2(), weight: 1e-1, rotWeight: 1e2))
-    }),
-  addFixedBetweenFactor: { (values, variables, graph) -> () in
-    let (prior) = unpack(values)
-    let (poseID) = unpack(variables)
-    graph.store(WeightedPriorFactorPose2SD(poseID, prior, sdX: 8, sdY: 4.6, sdTheta: 0.3))
-  })
-}
-
 /// Returns a tracking configuration for a tracker using PCA.
 ///
 /// Parameter model: The PCA model to use.
